requirement.txt
wikipedia
chatbotAI
python-amazon-simple-product-api